Changes in socket
Changed some useless things in socket and fixed group gate bugs in gamemap

+++ Oblivion/Oblivion.Emulator/Configuration/ConnectionHandling.cs
@@ -1,6 +1,7 @@
 using System;
 using Oblivion.Connection.Connection;
 using Oblivion.Connection.Net;
+using Oblivion.Util;
 
 namespace Oblivion.Configuration
 {
@@ -40,7 +41,8 @@ namespace Oblivion.Configuration
         {
             try
             {
-                Oblivion.GetGame().GetClientManager().CreateAndStartClient(connection.GetConnectionId(), connection);
+                Oblivion.GetGame().GetClientManager().CreateAndStartClient((uint)connection.GetConnectionId(), connection);
+
             }
             catch (Exception ex)
             {
@@ -56,7 +58,7 @@ namespace Oblivion.Configuration
         {
             try
             {
-                Oblivion.GetGame().GetClientManager().DisposeConnection(connection.GetConnectionId());
+                Oblivion.GetGame().GetClientManager().DisposeConnection((uint)connection.GetConnectionId());
             }
             catch (Exception ex)
             {
